requirements.txt
Django==3.2.14
asgiref
install
pytz
sqlparse
python-decouple
django-simple-captcha
django-multi-captcha-admin
django-simple-math-captcha==1.1.0
django-robots
mysqlclient